WEAR/requirements.txt
torch>=2.2
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
matplotlib>=3.7
pyyaml>=6.0
